Rename splice_pipeline.py to junction_pipeline.py
